% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical_mass_balance.R
\name{bf_chem_mass_balance}
\alias{bf_chem_mass_balance}
\title{Baseflow Separation by use of a chemical mass balance}
\usage{
bf_chem_mass_balance(date, discharge, C, C_baseflow = max(C),
  C_runoff = min(C), STAID = "Unknown")
}
\arguments{
\item{date}{vector of dates corresponding to each \code{discharge}, should be of class "Date."
Missing values are not permitted.}

\item{discharge}{the daily streamflow to be separated missing values are not permitted}

\item{C}{daily chemical concentration values}

\item{C_baseflow}{assumed chemical concentration associated with water derived from baseflow}

\item{C_runoff}{assumed chemical concentration associated with water deriver from surface flow}

\item{STAID}{the station identifier for the data.}
}
\value{
an object of class "baseflow" and inherits class "data.frame" of the selected data,
a data frame of the baseflow information, and other information about the analysis.
}
\description{
Extract baseflow from a daily streamflow record using conductance or other continuously
monitored conservative chemicals as markers of baseflow and surface flow.
}
\examples{

\dontrun{
}
}
\keyword{baseflow}
