% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eckhardt_digital_filter.R
\name{bf_eckhardt_filter}
\alias{bf_eckhardt_filter}
\title{Baseflow Separation by use of an Eckhardt Filter}
\usage{
bf_eckhardt_filter(date, discharge, BFImax, alpha, STAID = "Unknown")
}
\arguments{
\item{date}{vector of dates corresponding to each \code{discharge}, should be of class "Date."
Missing values are not permitted.}

\item{discharge}{the daily streamflow to be separated missing values are not permitted
within the time specified by \code{Start} and \code{end}.}

\item{BFImax}{maximum BFI to use in the filtering.}

\item{alpha}{filter parameter.}

\item{STAID}{the station identifier for the data.}
}
\value{
an object of class "baseflow" and inherits class "data.frame" of the selected data,
a data frame of the baseflow information, and other information about the analysis.
}
\description{
Extract baseflow from a daily streamflow record using the method described by
Eckhardt (2005).
}
\examples{

\dontrun{
}
}
\references{
Eckhardt, K., 2005, How to construct recursive digital filters for
 baseflow separation: Hydrological Processes, v. 19, no. 2, p. 507–515.
}
\keyword{baseflow}
