% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hysep.R
\name{bf_hysep}
\alias{bf_hysep}
\title{Baseflow Separation}
\usage{
bf_hysep(date, discharge, da, select = "sliding", STAID = "Unknown")
}
\arguments{
\item{date}{vector of dates for each \code{x}, should be of class "Date." Missing values
are not permitted.}

\item{discharge}{the daily streamflow to be separated missing values are not permitted
within the time specified by \code{Start} and \code{end}.}

\item{da}{the drainage area of the basin in square miles.}

\item{select}{a character string indicating which method to use for the
baseflow in the output dataset. Must be one of "sliding," "local minimum,"
or "fixed." Onle the first letter is required.}

\item{STAID}{the station identifier for the data.}
}
\value{
an object of class "baseflow" and inherits class "data.frame" of the selected data,
a data frame of the baseflow information, and other information about the analysis.
}
\description{
Extract baseflow from a daily streamflow record using the method described by
Sloto and Crouse (1996).
}
\examples{

\dontrun{
library(smwrData)
data(ChoptankFlow)
# Process by calendar year as that is the retrieval range
ChopPart <- with(ChoptankFlow, hysep(Flow, datetime, da=113,
STAID="01491000"))
ChopPart
}
}
\references{
Sloto, R.A. and Crouse, M.Y., 1996, HYSEP:
A COMPUTER PROGRAM FOR STREAMFLOW HYDROGRAPH SEPARATION AND ANALYSIS:
U.S. geological Survey Water-Resources Investigations
Report 96-4040. 46 p.
}
\keyword{baseflow}
