% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_term_digital_filter.R
\name{bf_single_term_filter}
\alias{bf_single_term_filter}
\title{Baseflow Separation by use of a single-term digital filter.}
\usage{
bf_single_term_filter(date, discharge, alpha, STAID = "Unknown")
}
\arguments{
\item{date}{vector of dates corresponding to each \code{discharge}, should be of class "Date."
Missing values are not permitted.}

\item{discharge}{the daily streamflow to be separated missing values are not permitted
within the time specified by \code{Start} and \code{end}.}

\item{alpha}{filter parameter.}

\item{STAID}{the station identifier for the data.}
}
\value{
an object of class "baseflow" and inherits class "data.frame" of the selected data,
a data frame of the baseflow information, and other information about the analysis.
}
\description{
Extract baseflow from a daily streamflow record using the method described by
Nathan and MacMahon (1990).
}
\examples{

\dontrun{
}
}
\keyword{baseflow}
