% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfi.R
\name{bfi}
\alias{bfi}
\title{Baseflow Separation}
\usage{
bfi(date, discharge, by = "water year", f = 0.9, N = 5L,
  STAID = "Unknown")
}
\arguments{
\item{date}{the date for each \code{discharge} value, should be of class "Date." Missing values
are not permitted.}

\item{discharge}{the daily streamflow to be separated missing values are not permitted
within the time specified by \code{Start} and \code{end}.}

\item{by}{string describing how to construct turning points: "water year", "calendar_year", or
"continuous".}

\item{f}{the factor for identifying turning points.}

\item{N}{the time perod for period for calculating the turning points.}

\item{STAID}{the station identifier for the data.}
}
\value{
an object of class "baseflow" and inherits class "data.frame" of the selected data, a data
frame of the baseflow information, and other information about the analysis.
}
\description{
Extract baseflow from a daily streamflow record using the method proposed by
the Institute of Hydrology (1980a,b), and as implemented by Wahl and Wahl (1988).
}
\details{
The original description of constructing turning points proposed dividing
water years into 5-day increments and selecting the minimum flow for each period.
Gustard and others (1992) describe using continuous 5-day increments rather than restarting
on the water year.
}
\examples{

\dontrun{
# Process by calendar year as that is the retrieval range
ChopBFI <- with(ChoptankFlow, bfi(date, discharge, by="calendar year",
STAID="01491000"))
head(ChopBFI, 20)
}
}
\references{
Gustard, A., Bullock, A., and Dixon, J.M., 1992,
Low flow estimation in the United Kingdom: Institue of Hydrology
Report No. 108, 88 p. and appendixes.

Institute of Hydrology, 1980a, Low flow studies:
Wallingford, Oxon, United Kingdom, Report No. I, 41 p.

Institute of Hydrology, 1980b, Low flow studies: Wallingford, Oxon,
United Kingdom, Report No. 3, p. 12- 19.

Wahl, K.L., and Wahl, T.L., 1988, BFI—A computer program for determining an index
to base flow: US Bureau of Reclamation Water Resources Research Laboratory,
 accessed August 25, 2004: U.S. Bureau of Reclamation.
}
\keyword{baseflow}
