% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfs_to_recharge.R
\name{cfs_to_recharge}
\alias{cfs_to_recharge}
\title{Convert baseflow to a recharge estimate for a basin}
\usage{
cfs_to_recharge(date, baseflow, da)
}
\arguments{
\item{date}{the date corresponding to the baseflow (class "Date").}

\item{baseflow}{the mean daily baseflow for the corresponding date.}

\item{da}{the drainage area of the basin in square miles.}
}
\value{
A data frame containing the following fields:
   date date corresponding to the baseflow value
   baseflow baseflow value in cfs
   month
   day
   year
   wy
   total_daily_recharge_ft3
   recharge_in
}
\description{
Convert the daily baseflow in cfs to a net infiltration (recharge) estimate
in inches over a basin.
}
