% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrobudget.R
\name{compute_hydrobudget}
\alias{compute_hydrobudget}
\title{Simulation using HydroBudget model}
\usage{
compute_hydrobudget(
  calibration,
  input_rcn,
  input_rcn_gauging,
  input_climate,
  simul_period,
  nb_core = NULL
)
}
\arguments{
\item{calibration}{The calibration parameters.}

\item{input_rcn}{The RCN values. Input can be a data.frame/data.table or a path to a data file.}

\item{input_rcn_gauging}{The table with the list of RCN cells located in each gauging station watershed. Input can be a data.frame/data.table or a path to a data file.}

\item{input_climate}{The daily total precipitation (mm/d) and average daily temperature (°C). Input can be a data.frame/data.table or a path to a data file.}

\item{simul_period}{The start and end years.}

\item{nb_core}{The number of cores to use in the parallel computations. If not provided, all cores minus one will be used.}
}
\value{
The water budget
}
\description{
HydroBudget is a spatially distributed GWR model that computes a superficial water budget on
grid cells of regional-scale watersheds. Runoff, actual evapotranspiration (AET), and potential
GWR are simulated for each grid cell, with a monthly time step, and fluxes do not
transfer from a cell to another (no water routing). The model inputs are distributed daily
precipitation and temperature as well as distributed data of pedology, land cover, and slope.
}
