% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrobudget-simulation.R
\name{compute_potential_evapotranspiration_cell}
\alias{compute_potential_evapotranspiration_cell}
\title{Compute PET based on the Oudin formula}
\usage{
compute_potential_evapotranspiration_cell(input_dd)
}
\arguments{
\item{input_dd}{Spatially distributed daily precipitation and mean temperature time series in a single cell.}
}
\value{
Spatially distributed daily PET
}
\description{
Compute PET based on the Oudin formula
}
\keyword{internal}
