% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{compute_simulation_quality_assessment}
\alias{compute_simulation_quality_assessment}
\title{Simulation quality assessment}
\usage{
compute_simulation_quality_assessment(
  calibration,
  water_budget,
  input_rcn_gauging,
  observed_flow,
  alpha_lyne_hollick,
  simul_period
)
}
\arguments{
\item{calibration}{The calibration parameters.}

\item{water_budget}{The computed water budget.}

\item{input_rcn_gauging}{The table with the list of RCN cells located in each gauging station watershed. Input can be a data.frame/data.table or a path to a data file.}

\item{observed_flow}{The flow rates in mm/day. Input can be a data.frame/data.table or a path to a data file.}

\item{alpha_lyne_hollick}{The Lyne and Hollick filter. Input can be a data.frame/data.table or a path to a data file.}

\item{simul_period}{The start and end years.}
}
\description{
Simulation quality assessment
}
