% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrobudget.R
\name{compute_vertical_inflow}
\alias{compute_vertical_inflow}
\title{Determine if precipitation is rain or snow and simulate the snowpack (accumulation
and melt) to compute the vertical inflow (VI), the liquid water available per day (rainfall + melt water).}
\usage{
compute_vertical_inflow(climate_data, T_m, C_m, T_snow, nb_core)
}
\description{
Determine if precipitation is rain or snow and simulate the snowpack (accumulation
and melt) to compute the vertical inflow (VI), the liquid water available per day (rainfall + melt water).
}
\keyword{internal}
