% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrobudget.R
\name{compute_vertical_inflow_cell}
\alias{compute_vertical_inflow_cell}
\title{Compute the vertical inflow and the potential evapotranspiration (PET) for a single cell}
\usage{
compute_vertical_inflow_cell(calibration, input_dd)
}
\arguments{
\item{calibration}{The calibration parameters.}

\item{input_dd}{Spatially distributed daily precipitation and mean temperature time series in a single cell.}
}
\description{
Compute the vertical inflow and the potential evapotranspiration (PET) for a single cell
}
\keyword{internal}
