% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrobudget.R
\name{compute_water_budget}
\alias{compute_water_budget}
\title{Compute the water budget}
\usage{
compute_water_budget(calibration, climate_data, rcn, nb_core)
}
\arguments{
\item{calibration}{The calibration parameters.}

\item{climate_data}{The daily total precipitation (mm/d) and average daily temperature (°C).}

\item{rcn}{The RCN values.}

\item{nb_core}{The number of cores to use in the parallel computations.}
}
\description{
Compute the water budget
}
\keyword{internal}
