% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrobudget-simulation.R
\name{compute_water_budget_cell}
\alias{compute_water_budget_cell}
\title{Compute the water budget for a single cell}
\usage{
compute_water_budget_cell(obj, rcn_climate)
}
\arguments{
\item{obj}{The HydroBudget object.}

\item{rcn_climate}{The RCN values with the climate data (daily total precipitation (mm/d) and average daily temperature (°C)) for a single cell.}
}
\description{
Compute the water budget for a single cell
}
\keyword{internal}
