% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.as.data.table}
\alias{.as.data.table}
\title{Make a data.table from input: coerce it or read it (if it is a file path)}
\usage{
.as.data.table(input)
}
\description{
Make a data.table from input: coerce it or read it (if it is a file path)
}
\keyword{internal}
