% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is.verbose}
\alias{.is.verbose}
\title{Make a data.table from input: coerce it or read it (if it is a file path) and ensure column names}
\usage{
.is.verbose()
}
\description{
Make a data.table from input: coerce it or read it (if it is a file path) and ensure column names
}
\keyword{internal}
