% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.newProgress}
\alias{.newProgress}
\title{Create a new progress instance with default settings.}
\usage{
.newProgress(
  format = "  :what [:bar] :current/:total | :elapsed",
  clear = getOption("recharge.progress.clear", !interactive()),
  total,
  width = 100
)
}
\description{
Create a new progress instance with default settings.
}
\keyword{internal}
