% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.tickProgress}
\alias{.tickProgress}
\title{Update and increment the progress status if option "recharge.progress" is TRUE.}
\usage{
.tickProgress(progress, tokens = list())
}
\description{
Update and increment the progress status if option "recharge.progress" is TRUE.
}
\keyword{internal}
