% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.updateProgress}
\alias{.updateProgress}
\title{Update the progress status if option "recharge.progress" is TRUE.}
\usage{
.updateProgress(progress, step, total, tokens = list())
}
\description{
Update the progress status if option "recharge.progress" is TRUE.
}
\keyword{internal}
