% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_os.R
\name{get_os}
\alias{get_os}
\title{get_os}
\usage{
get_os()
}
\description{
Determine what operating system R is running under.
}
\details{
This code is adapted from a post discussing a safe way to
detect the operating system that R is running under:
https://www.r-bloggers.com/identifying-the-os-from-r/

Specifically, this code is designed to return the proper operating
system name even if the R function 'Sys.info' is unimplemented.
}
