% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KGE.R
\name{KGE}
\alias{KGE}
\title{Kling-Gupta Efficiency computation}
\usage{
KGE(sim, obs)
}
\arguments{
\item{sim}{simulated values}

\item{obs}{observed values}
}
\value{
Kling-Gupta Efficiency between 'sim' and 'obs'
}
\description{
Kling-Gupta Efficiency computation
}
\keyword{internal}
