% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrobudget.R
\name{make_calibration_parameters}
\alias{make_calibration_parameters}
\title{Validate calibration parameters}
\usage{
make_calibration_parameters(calibration = list())
}
\description{
Validate calibration parameters
}
\keyword{internal}
