% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrobudget-simulation.R
\name{new_hydrobugdet}
\alias{new_hydrobugdet}
\title{HydroBudget object}
\usage{
new_hydrobugdet(T_m, C_m, TT_F, F_T, t_API, f_runoff, sw_m, f_inf)
}
\arguments{
\item{T_m}{The melting temperature (°C)}

\item{C_m}{The melting coefficient (mm/°C/d)}

\item{TT_F}{The Threshold temperature for soil frost (°C)}

\item{F_T}{The freezing time (d)}

\item{t_API}{The antecedent precipitation index time (d)}

\item{f_runoff}{The runoff factor (-)}

\item{sw_m}{The maximum soil water content (mm)}

\item{f_inf}{The infiltration factor (-)}
}
\description{
Make a new HydroBudget object, by providing the calibration parameters for the
model computation.
}
