% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_river_flow}
\alias{process_river_flow}
\title{River flow processing}
\usage{
process_river_flow(observed_flow, alpha_lyne_hollick)
}
\arguments{
\item{observed_flow}{The flow rates in mm/day.}

\item{alpha_lyne_hollick}{The Lyne and Hollick filter.}
}
\value{
A list of observed_flow_month and gauging
}
\description{
Processing of the river flow to compute the baseflow with Lyne and Hollick and resample them into monthly observations by
(1) select the available river flow observations for the simulation period and fill the gaps (up to 5 missing days),
(2) extract the list of the available gauging stations for the simulation period (list of the names),
(3) compute the Lyne and Hollick baseflow and resample river flow and baseflow with a monthly time step.
}
