% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recharge-package.R
\docType{package}
\name{recharge-package}
\alias{recharge-package}
\alias{recharge}
\title{Functions for estimating groundwater recharge from streamflow data}
\description{
This package contains several algorithms designed to estimate groundwater
recharge (net infiltration) from daily streamflow data.
}
\details{
\tabular{ll}{ Package: \tab recharge\cr
Type: \tab Package\cr
License: \tab CC0\cr
Depends: \tab lubridate, dplyr\cr }
This package contains functions that manage or manipipulate
hydrologic daily-value data. The functions in this package
are listed below, grouped by a general description of the task.

Functions to compute baseflow from the streamflow record and
functions that summarize those computations.\cr
\code{\link{bfi}}\cr
\code{\link{hysep}}\cr
\code{\link{part}}\cr
\code{\link{aggregate.baseflow}}\cr
\code{\link{plot.baseflow}}\cr
\code{\link{print.baseflow}}\cr

Functions that analyze groundwater record, primarily for the
estimation of recharge.\cr
\code{\link{fall}}\cr
\code{\link{rise}}\cr
\code{\link{wtf}}\cr
\code{\link{aggregate.rise}}\cr
\code{\link{confirm.fall}}\cr
\code{\link{plot.fall}}\cr
\code{\link{plot.rise}}\cr
\code{\link{print.fall}}\cr
\code{\link{print.rise}}\cr
\code{subset.fall}\cr

Functions for streamflow recession analysis.\cr
\code{\link{join}}\cr
\code{\link{recess}}\cr
\code{\link{rora}}\cr
\code{\link{aggregate.rora}}\cr
\code{\link{confirm.recess}}\cr
\code{\link{plot.recess}}\cr
\code{\link{plot.rora}}\cr
\code{\link{print.recess}}\cr
}
\references{
Need those for the various routines!
}
\seealso{
\code{\link[smwrData:smwrData-package]{smwrData}}
}
\author{
Steve Westenbroek
}
\keyword{package}
