% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{with_progress}
\alias{with_progress}
\title{Progress option}
\usage{
with_progress(progress = TRUE)
}
\arguments{
\item{progress}{Logical to set for having a progress bar}
}
\value{
(Invisible) the return value of \link[progressr]{handlers}
}
\description{
Progress option
}
