% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{with_verbose}
\alias{with_verbose}
\title{Verbose option}
\usage{
with_verbose(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical to set for having verbose messages}
}
\value{
(Invisible) the return value of \link[base]{options}
}
\description{
Verbose option
}
