% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{write_rasters}
\alias{write_rasters}
\title{Write result as raster files}
\usage{
write_rasters(water_budget, input_rcn, crs, output_dir = getwd())
}
\arguments{
\item{water_budget}{The computed water budget. Input can be a data.frame/data.table or a path to a data file.}

\item{input_rcn}{The RCN values. Input can be a data.frame/data.table or a path to a data file.}

\item{crs}{The coordinate reference systems.}

\item{output_dir}{The output directory where result files will be written. Default is current working directory.}
}
\description{
Export raster for interannual runoff, aet and GWR.
}
