% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/hydrobudget-results.R
\name{write_recharge_results}
\alias{write_recharge_results}
\alias{write_recharge_results.default}
\alias{write_recharge_results.hydrobudget}
\title{Write result as data files}
\usage{
write_recharge_results(obj, water_budget, output_dir = getwd(), ...)

\method{write_recharge_results}{default}(obj, water_budget, output_dir = getwd(), ...)

\method{write_recharge_results}{hydrobudget}(obj, water_budget, output_dir = getwd(), ...)
}
\arguments{
\item{obj}{The recharge object.}

\item{water_budget}{The computed water budget.}

\item{output_dir}{The output directory where result files will be written. Default is current working directory.}

\item{...}{Other arguments passed to methods}
}
\description{
Export water budget.
}
