% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrobudget.R
\name{write_results}
\alias{write_results}
\title{Write result as data files}
\usage{
write_results(water_budget, output_dir = getwd())
}
\arguments{
\item{water_budget}{The computed water budget.}

\item{output_dir}{The output directory where result files will be written. Default is current working directory.}
}
\description{
Export water budget.
}
