% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrobudget.R
\name{compute_potential_evapotranspiration_cell}
\alias{compute_potential_evapotranspiration_cell}
\title{Compute potential evapotranspiration (PET) based on the Oudin formula for a single cell.}
\usage{
compute_potential_evapotranspiration_cell(cellgrid, input_dd, T_m, C_m, T_snow)
}
\description{
Compute potential evapotranspiration (PET) based on the Oudin formula for a single cell.
}
\keyword{internal}
