% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrobudget-quality.R
\name{compute_simulation_quality_assessment}
\alias{compute_simulation_quality_assessment}
\title{Simulation quality assessment}
\usage{
compute_simulation_quality_assessment(
  obj,
  water_budget,
  rcn_gauging,
  observed_flow,
  alpha_lyne_hollick,
  period = NULL
)
}
\arguments{
\item{obj}{The HydroBudget object with calibration parameters and column names mappings.}

\item{water_budget}{The computed water budget. Input can be a data.frame/data.table or a path to a data file.}

\item{rcn_gauging}{The table with the list of RCN cells located in each gauging station watershed. Input can be a data.frame/data.table or a path to a data file.}

\item{observed_flow}{The flow rates in mm/day. Input can be a data.frame/data.table or a path to a data file.}

\item{alpha_lyne_hollick}{The Lyne and Hollick filter. Input can be a data.frame/data.table or a path to a data file.}

\item{period}{The start and end years. If not provided, the start/end years will be extracted from the water budget data.}
}
\description{
Evaluates the simulated water budget with the average KGE.
}
\details{
The columns of the water budget data set input are:
\itemize{
\item \strong{year}
\item \strong{month}
\item \strong{VI}
\item \strong{t_mean}
\item \strong{runoff}
\item \strong{pet}
\item \strong{aet}
\item \strong{gwr}
\item \strong{runoff_2}
\item \strong{delta_reservoir}
\item \strong{rcn_id}
}

The columns of the observed flow data set input are:
\itemize{
\item \strong{year}
\item \strong{month}
\item \strong{day}
\item \strong{one column per station} (named by the station ID), the flow rates in mm/day
}

The columns of the RCN gauging stations data set input are:
\itemize{
\item \strong{rcn_id}, the cell ID
\item \strong{station_id}, the station ID
}

The columns of the Lyne and Hollick filter data set input are:
\itemize{
\item \strong{station_id}, the station ID
\item \strong{alpha}
}
}
\examples{
\dontrun{
# Use input example files provided by the package
examples_dir <- system.file("examples", package = "rechaRge")
observed_flow <- file.path(examples_dir, "input", "observed_flow.csv.gz") # flow rates in mm/d
alpha_lyne_hollick <- file.path(examples_dir, "input", "alpha_lyne_hollick.csv.gz")

# Calibration parameters
HB <- rechaRge::new_hydrobudget(
  T_m = 2.1, # melting temperature (°C)
  C_m = 6.2, # melting coefficient (mm/°C/d)
  TT_F = -17.6, # Threshold temperature for soil frost (°C)
  F_T = 16.4, # Freezing time (d)
  t_API = 3.9, # Antecedent precipitation index time (d)
  f_runoff = 0.63, # Runoff factor (-)
  sw_m = 431, # Maximum soil water content (mm)
  f_inf = 0.07 # infiltration factor (-)
)

# ... compute the water budget ...

result <- compute_simulation_quality_assessment(
  HB,
  water_budget = water_budget,
  rcn_gauging = input_rcn_gauging,
  observed_flow = observed_flow,
  alpha_lyne_hollick = alpha_lyne_hollick,
  period = simul_period
)
}
}
