% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrobudget.R
\name{compute_vertical_inflow}
\alias{compute_vertical_inflow}
\title{Determine if precipitation is rain or snow and simulate the snowpack (accumulation
and melt) to compute the vertical inflow (VI), the liquid water available per day (rainfall + melt water).}
\usage{
compute_vertical_inflow(calibration, climate_data, nb_core)
}
\arguments{
\item{calibration}{The calibration parameters.}

\item{climate_data}{The daily total precipitation (mm/d) and average daily temperature (°C).}

\item{nb_core}{The number of cores to use in the parallel computations.}
}
\description{
Determine if precipitation is rain or snow and simulate the snowpack (accumulation
and melt) to compute the vertical inflow (VI), the liquid water available per day (rainfall + melt water).
}
\keyword{internal}
