% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrobudget-simulation.R
\name{compute_vertical_inflow}
\alias{compute_vertical_inflow}
\title{Determine if precipitation is rain or snow and simulate the snowpack (accumulation
and melt) to compute the vertical inflow (vi), the liquid water available per day (rainfall + melt water).}
\usage{
compute_vertical_inflow(obj, climate_data)
}
\arguments{
\item{obj}{The HydroBudget object.}

\item{climate_data}{The daily total precipitation (mm/d) and average daily temperature (°C).}
}
\description{
Determine if precipitation is rain or snow and simulate the snowpack (accumulation
and melt) to compute the vertical inflow (vi), the liquid water available per day (rainfall + melt water).
}
\keyword{internal}
