% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrobudget-simulation.R
\name{compute_water_budget}
\alias{compute_water_budget}
\title{Compute the water budget}
\usage{
compute_water_budget(obj, rcn_data, climate_data, workers)
}
\arguments{
\item{obj}{The HydroBudget object.}

\item{rcn_data}{The RCN values.}

\item{climate_data}{The daily total precipitation (mm/d) and average daily temperature (°C).}

\item{workers}{The number of workers to use in the parallel computations.}
}
\description{
Compute the water budget
}
\keyword{internal}
