% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrc.R
\name{mrc}
\alias{mrc}
\title{mrc}
\usage{
mrc(x, time_var = "datetime", precip_var = "precip", n = 48,
  lag = 48, cutoff = 0.01)
}
\arguments{
\item{lag}{A vector of positive integers. Each specified column will be
lagged for each value in the vector.}
}
\value{
data.table with mrc
}
\description{
mrc
}
